requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
motor==3.5.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
pillow==10.4.0
numpy==2.0.2
scikit-learn==1.5.1
openai==1.40.6
